Improves index and links to JS instead of include
As I noticed, index.html and repo's README.md are not synchronised. So I added links to ToDo.
Noticed, that I've used a reltive link to the JS that only works on the repo so last time I included the JS.
But now it's removed again and linked to latest repo-version instead.